Attempting to figure out ComboBox.kt
In this commit, I have been trying to get ComboBox.kt to work properly thus that one can create on in kotlin purely. This as been frustrating as the result always ends up with a blank spinner.

I am considering creating the entire demo in kotlin to do a comparison

- build.gradle.kts    : New combobox test
- ComboBox.kt         : Created DSL for ComboBox
- combobox.kt         : Created minor part from gtk-demo
- Converters.kt       : Array<String -> Int>
- FrameClock.kt       : Added annotations to shut up compiler
- KGType.kt           : Basic wrapper to test with
- KGValue.kt          : dsl/nativex Basic wrapper to test with
- ListStore.kt        : dsl/nativex Basic wrapper class to test
- Main.kt             : Annotations added
- settings.gradle.kts : Added new sub project
- Signal.kt           : Removed space
- TreeModel.kt        : Spacing fixed & constructor for Iter
- UI.kt               : Annotations added